+++ packages/map-template/src/atoms/apiKeyState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const apiKeyState = atom({
+    key: 'apiKey',
+    default: null
+});
+
+export default apiKeyState;