+++ packages/map-template/src/atoms/currentLocationState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const currentLocationState = atom({
+    key: 'currentLocation',
+    default: null
+});
+
+export default currentLocationState;