+++ packages/map-template/src/atoms/directionsServiceState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const directionsServiceState = atom({
+    key: 'directionsService',
+    default: null
+});
+
+export default directionsServiceState;