+++ packages/map-template/src/atoms/userPositionState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const userPositionState = atom({
+    key: 'userPosition',
+    default: null
+});
+
+export default userPositionState;