+++ packages/map-template/src/atoms/venuesState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const venuesState = atom({
+    key: 'venues',
+    default: []
+});
+
+export default venuesState;