+++ packages/map-template/src/components/Directions/Directions.jsx
@@ -1,6 +1,8 @@
-import React, { useContext, useEffect, useState, useRef } from "react";
+import React, { useEffect, useState, useRef } from "react";
 import './Directions.scss';
-import { MapsIndoorsContext } from '../../MapsIndoorsContext';
+import { useRecoilValue } from 'recoil';
+import mapsIndoorsInstanceState from '../../atoms/mapsIndoorsInstanceState';
+import travelModeState from '../../atoms/travelModeState';
 import { ReactComponent as CloseIcon } from '../../assets/close.svg';
 import { ReactComponent as ClockIcon } from '../../assets/clock.svg';
 import { ReactComponent as WalkingIcon } from '../../assets/walk.svg';
@@ -21,9 +23,8 @@ let directionsRenderer;
  * @param {boolean} props.isOpen - Indicates if the directions view is open.
  * @param {function} props.onBack - Callback that fires when the directions view is closed by the user.
  * @param {function} props.directions - The directions information based on the origin and destination.
- * @param {string} props.selectedTravelMode - The selected travel mode chosen by the user.
  */
-function Directions({ isOpen, onBack, directions, selectedTravelMode }) {
+function Directions({ isOpen, onBack, directions }) {
     // Holds the MapsIndoors DisplayRule for the destination
     const [destinationDisplayRule, setDestinationDisplayRule] = useState(null);
 
@@ -33,7 +34,8 @@ function Directions({ isOpen, onBack, directions, selectedTravelMode }) {
     const [totalDistance, setTotalDistance] = useState();
     const [totalTime, setTotalTime] = useState();
 
-    const mapsIndoorsInstance = useContext(MapsIndoorsContext);
+    const mapsIndoorsInstance = useRecoilValue(mapsIndoorsInstanceState);
+    const travelMode = useRecoilValue(travelModeState);
 
     const isDesktop = useMediaQuery('(min-width: 992px)');
 
@@ -72,7 +74,7 @@ function Directions({ isOpen, onBack, directions, selectedTravelMode }) {
                 setDestinationDisplayRule(mapsIndoorsInstance.getDisplayRule(directions.destinationLocation));
             }
         }
-    }, [isOpen, directions, mapsIndoorsInstance, selectedTravelMode]);
+    }, [isOpen, directions, mapsIndoorsInstance, travelMode]);
 
     /*
      * Make sure directions stop rendering on the map when the Directions view is not active anymore.
@@ -192,7 +194,7 @@ function Directions({ isOpen, onBack, directions, selectedTravelMode }) {
                                     <div className='directions__name'>
                                         {directions?.destinationLocation.properties.name}
                                     </div>
-                                    {directions?.originLocation.properties.name !== 'My Position' && <mi-location-info ref={destinationInfoElement} />}
+                                    <mi-location-info ref={destinationInfoElement} />
                                 </div>
                             </div>
                         }
@@ -202,9 +204,9 @@ function Directions({ isOpen, onBack, directions, selectedTravelMode }) {
             <div className="directions__guide">
                 <div className="directions__metrics">
                     <div className="directions__distance">
-                        {selectedTravelMode === travelModes.WALKING && <WalkingIcon />}
-                        {selectedTravelMode === travelModes.DRIVING && <DriveIcon />}
-                        {selectedTravelMode === travelModes.BICYCLING && <BikeIcon />}
+                        {travelMode === travelModes.WALKING && <WalkingIcon />}
+                        {travelMode === travelModes.DRIVING && <DriveIcon />}
+                        {travelMode === travelModes.BICYCLING && <BikeIcon />}
                         <div>Distance:</div>
                         <div className="directions__meters">{totalDistance && <mi-distance meters={totalDistance} />}</div>
                     </div>