+++ packages/map-template/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2023-06-12
+
+### Added
+
+- The Map Template now uses [Recoil](https://recoiljs.org/) to manage state.
+
+### Fixed
+
+- Fix crash in the Wayfinding component when having `My Position` enabled.
+
+## [1.1.1] - 2023-06-08
+
+### Fixed
+
+- Fix missing prop for `startZoomLevel`.
+
 ## [1.1.0] - 2023-06-08
 
 ### Added